requirements.txt
# --- Core Framework (Stable Versions) ---
streamlit==1.31.0
streamlit-drawable-canvas==0.9.3
Pillow>=10.2.0

# --- Document & Image Processing ---
pymupdf>=1.24.0
opencv-python>=4.9.0
numpy<2.0.0

# --- AI & Clustering ---
scikit-learn>=1.4.0
yomitoku==0.10.3

# --- Computation & Data ---
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.2.0
